+++ src/plugins/grpc/channelCredentialsRequestHook.ts
@@ -1,4 +1,4 @@
-import { ChannelCredentials } from '@grpc/grpc-js';
+import { ChannelCredentials, credentials } from '@grpc/grpc-js';
 import { HookCancel } from 'hookpoint';
 
 import * as models from '../../models';
@@ -8,6 +8,14 @@ import { isGrpcRequest } from './grpcRequest';
 export async function channelCredentialsRequestHook(request: models.Request): Promise<void | typeof HookCancel> {
   if (isGrpcRequest(request) && request.headers) {
     const channelCredentials = utils.getHeader(request.headers, 'channelcredentials');
+    if (utils.isString(channelCredentials)) {
+      if (channelCredentials.toLowerCase() === 'ssl') {
+        request.channelCredentials = credentials.createSsl();
+      }
+      if (channelCredentials.toLowerCase() === 'insecure') {
+        request.channelCredentials = credentials.createInsecure();
+      }
+    }
     if (channelCredentials instanceof ChannelCredentials) {
       utils.deleteHeader(request.headers, 'channelcredentials');
       request.channelCredentials = channelCredentials;